backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
langgraph
python-docx
requests
python-multipart
langchain-core
langchain-google-genai
langchain-groq
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
lxml
orjson
ormsgpack
anyio
httpx
pydantic
sqlalchemy
typing-extensions
typing-inspection
annotated-doc
annotated-types
greenlet
click
h11
starlette
requests-toolbelt
pyyaml
tenacity
uuid-utils
zstandard
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
numpy
scikit-learn
rank-bm25
spacy
